requirements.txt
mcp>=0.9.0
httpx>=0.25.0
python-dateutil>=2.8.2
pytest>=7.0.0
pytest-asyncio>=0.21.0
rich>=13.0.0
click>=8.1.0
ollama==0.5.3
python-dotenv==1.1.1
starlette==0.47.2
